+++ server-plugin/Code/Systems/Blockers/BhopBlocker.cpp
@@ -132,7 +132,7 @@ PlayerRunCommandRet BhopBlocker::RT_PlayerRunCommandCallback ( PlayerHandler::co
 	float const tick_interval ( SourceSdk::InterfacesProxy::Call_GetTickInterval () );
 	__assume ( tick_interval > 0.0f && tick_interval < 1.0f );
 
-	int const blocking_delay_ticks ( floorf ( blocking_delay_seconds / tick_interval ) );
+	int const blocking_delay_ticks ( (int)floorf ( blocking_delay_seconds / tick_interval ) );
 
 	int const gtick ( Helpers::GetGameTickCount () );
